+++ src/index.ts
@@ -1,3 +1,3 @@
-export { default as loadData } from "./methods/loadData.js"
+export { default as loadData } from "./functions/loadData.js"
 
 export { default as SimpleData } from "./class/SimpleData.js"
