+++ src/methods/removeKey.ts
@@ -4,18 +4,6 @@ import showTable from "./showTable.js"
 
 export default function removeKey(data: SimpleDataItem[], key: string, options: Options): SimpleDataItem[] {
 
-    const start = Date.now()
-
-    options = {
-        ...defaultOptions,
-        ...options
-    }
-
-    options.logs && log("\nremoveKey() " + key)
-    options.logOptions && log("options:")
-    options.logOptions && log(options)
-
-    // All items needs to have the same keys
     if (!data[0].hasOwnProperty(key)) {
         throw new Error("No key " + key)
     }
@@ -23,11 +11,6 @@ export default function removeKey(data: SimpleDataItem[], key: string, options: 
     for (let i = 0; i < data.length; i++) {
         delete data[i][key]
     }
-
-    options.logs && showTable(data, options)
-
-    const end = Date.now()
-    options.logs && log(`Done in ${((end - start) / 1000).toFixed(3)} sec.`)
 
     return data
 }
