+++ src/methods/renameKey.ts
@@ -4,18 +4,6 @@ import showTable from "./showTable.js"
 
 export default function renameKey(data: SimpleDataItem[], oldKey: string, newKey: string, options: Options): SimpleDataItem[] {
 
-    const start = Date.now()
-
-    options = {
-        ...defaultOptions,
-        ...options
-    }
-
-    options.logs && log("\nrenameKey() " + oldKey + " " + newKey)
-    options.logOptions && log("options:")
-    options.logOptions && log(options)
-
-    // All items needs to have the same keys
     if (!data[0].hasOwnProperty(oldKey)) {
         throw new Error("No key " + oldKey)
     }
@@ -25,11 +13,6 @@ export default function renameKey(data: SimpleDataItem[], oldKey: string, newKey
         d[newKey] = d[oldKey]
         delete d[oldKey]
     }
-
-    options.logs && showTable(data, options)
-
-    const end = Date.now()
-    options.logs && log(`Done in ${((end - start) / 1000).toFixed(3)} sec.`)
 
     return data
 }
