+++ src/methods/roundValues.ts
@@ -1,21 +1,7 @@
-import log from "../helpers/log.js"
-import { SimpleDataItem, Options, defaultOptions } from "../types.js"
-import showTable from "./showTable.js"
+import { SimpleDataItem, Options } from "../types.js"
 
 export default function roundValues(data: SimpleDataItem[], key: string, options: Options): SimpleDataItem[] {
 
-    const start = Date.now()
-
-    options = {
-        ...defaultOptions,
-        ...options
-    }
-
-    options.logs && log("\nroundValues() " + key)
-    options.logOptions && log("options:")
-    options.logOptions && log(options)
-
-    // All items needs to have the same keys
     if (!data[0].hasOwnProperty(key)) {
         throw new Error("No key " + key)
     }
@@ -24,11 +10,6 @@ export default function roundValues(data: SimpleDataItem[], key: string, options
         //@ts-ignore
         data[i][key] = parseFloat(data[i][key].toFixed(options.fractionDigits))
     }
-
-    options.logs && showTable(data, options)
-
-    const end = Date.now()
-    options.logs && log(`Done in ${((end - start) / 1000).toFixed(3)} sec.`)
 
     return data
 }
