+++ src/methods/valuesToDate.ts
@@ -6,18 +6,6 @@ import { utcParse } from "d3-time-format"
 
 export default function valuesToDate(data: SimpleDataItem[], key: string, format: string, options: Options): SimpleDataItem[] {
 
-    const start = Date.now()
-
-    options = {
-        ...defaultOptions,
-        ...options
-    }
-
-    options.logs && log("\nvaluesToDate() " + key + " " + format)
-    options.logOptions && log("options:")
-    options.logOptions && log(options)
-
-    // All items needs to have the same keys
     if (!data[0].hasOwnProperty(key)) {
         throw new Error("No key " + key)
     }
@@ -28,11 +16,6 @@ export default function valuesToDate(data: SimpleDataItem[], key: string, format
         //@ts-ignore
         data[i][key] = parse(data[i][key])
     }
-
-    options.logs && showTable(data, options)
-
-    const end = Date.now()
-    options.logs && log(`Done in ${((end - start) / 1000).toFixed(3)} sec.`)
 
     return data
 }
