+++ src/methods/valuesToInteger.ts
@@ -1,21 +1,7 @@
-import log from "../helpers/log.js"
-import { SimpleDataItem, Options, defaultOptions } from "../types.js"
-import showTable from "./showTable.js"
+import { SimpleDataItem, Options } from "../types.js"
 
 export default function valuesToInteger(data: SimpleDataItem[], key: string, options: Options): SimpleDataItem[] {
 
-    const start = Date.now()
-
-    options = {
-        ...defaultOptions,
-        ...options
-    }
-
-    options.logs && log("\nvaluesToInteger() " + key)
-    options.logOptions && log("options:")
-    options.logOptions && log(options)
-
-    // All items needs to have the same keys
     if (!data[0].hasOwnProperty(key)) {
         throw new Error("No key " + key)
     }
@@ -24,11 +10,6 @@ export default function valuesToInteger(data: SimpleDataItem[], key: string, opt
         //@ts-ignore
         data[i][key] = parseInt(data[i][key])
     }
-
-    options.logs && showTable(data, options)
-
-    const end = Date.now()
-    options.logs && log(`Done in ${((end - start) / 1000).toFixed(3)} sec.`)
 
     return data
 }
